+++ test/controllers/internal_controller_test.rb
@@ -4,6 +4,7 @@ require 'test_helper'
 class InternalControllerTest < ActionController::TestCase
   setup do
     @viewer = users(:viewer)
+    @generic_viewer = viewers(:one)
   end
 
   test 'should get dashboard as viewer' do
@@ -23,6 +24,12 @@ class InternalControllerTest < ActionController::TestCase
     assert_response :success
   end
 
+  test 'should get directory as generic viewer' do
+    login_viewer(@generic_viewer)
+    get :directory
+    assert_response :success
+  end
+
   test 'should not get directory as anonymous user' do
     get :directory
     assert_redirected_to new_user_session_path
@@ -30,6 +37,13 @@ class InternalControllerTest < ActionController::TestCase
 
   test 'should get category as viewer' do
     login(@viewer)
+    get :category, category: categories(:one).slug
+    assert_not_nil assigns(:category)
+    assert_response :success
+  end
+
+  test 'should get category as generic viewer' do
+    login_viewer(@generic_viewer)
     get :category, category: categories(:one).slug
     assert_not_nil assigns(:category)
     assert_response :success
@@ -43,6 +57,16 @@ class InternalControllerTest < ActionController::TestCase
 
   test 'should download document as viewer' do
     login(@viewer)
+    get :document, category: categories(:one).slug, document_id: documents(:two)
+    assert_not_nil assigns(:category)
+    assert_not_nil assigns(:document)
+    assert_kind_of String, response.body
+    assert_equal File.binread(File.join(CarrierWave::Uploader::Base.root, assigns(:document).document.url)), response.body
+    assert_response :success
+  end
+
+  test 'should download document as generic viewer' do
+    login_viewer(@generic_viewer)
     get :document, category: categories(:one).slug, document_id: documents(:two)
     assert_not_nil assigns(:category)
     assert_not_nil assigns(:document)